refactor: countItem 함수 model에서 체크하도록 코드변경
